move configurator class into another file
