+++ src/index.ts
@@ -1,8 +1,7 @@
 import { ExternalService as InternalExternalService } from './metadata-decorators/external-service.decorator';
 import { HttpExceptionFilter } from './http-exception.filter';
 import ResponseNormalizerOptions from './configuration/response-normalizer.options';
-import { INestApplication } from '@nestjs/common';
-import { NormalizerInterceptor } from 'normalizer.interceptor';
+import { Configurator } from 'configuration/configurator';
 
 export const ExternalService = InternalExternalService;
 
@@ -13,18 +12,3 @@ export const init = function (
   nestApp.useGlobalFilters(new HttpExceptionFilter());
   new Configurator(nestApp, options);
 };
-
-export class Configurator {
-  public static options: ResponseNormalizerOptions;
-
-  constructor(nestApp: INestApplication, options?: ResponseNormalizerOptions) {
-    if (options === undefined) options = new ResponseNormalizerOptions();
-    Configurator.options = options;
-    this.init(nestApp);
-  }
-
-  private init(nestApp: INestApplication) {
-    if (Configurator.options.useNormalizerInterceptor)
-      nestApp.useGlobalInterceptors(new NormalizerInterceptor());
-  }
-}
